Mostly Fixed Movement stutter
It turned out to be related to the rigid body itself and not the networking.
I turned on interpolation and that seems to have fixed it for the most part.
I also seperated the HUD prefab from the Ship prefab, just need to do the same with the camera next